// src/Frontend/src/client/models/TimeSeriesDataWithProps.ts
/* tslint:disable */
/* eslint-disable */
/**
 * QueryApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    AdditionalTimeSeriesProperties,
    AdditionalTimeSeriesPropertiesFromJSON,
    AdditionalTimeSeriesPropertiesFromJSONTyped,
    AdditionalTimeSeriesPropertiesToJSON,
} from './AdditionalTimeSeriesProperties';
import {
    EventDataSet,
    EventDataSetFromJSON,
    EventDataSetFromJSONTyped,
    EventDataSetToJSON,
} from './EventDataSet';

/**
 * 
 * @export
 * @interface TimeSeriesDataWithProps
 */
export interface TimeSeriesDataWithProps {
    /**
     * 
     * @type {EventDataSet}
     * @memberof TimeSeriesDataWithProps
     */
    eventData?: EventDataSet;
    /**
     * 
     * @type {AdditionalTimeSeriesProperties}
     * @memberof TimeSeriesDataWithProps
     */
    additionalProps?: AdditionalTimeSeriesProperties;
}

export function TimeSeriesDataWithPropsFromJSON(json: any): TimeSeriesDataWithProps {
    return TimeSeriesDataWithPropsFromJSONTyped(json, false);
}

export function TimeSeriesDataWithPropsFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimeSeriesDataWithProps {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'eventData': !exists(json, 'eventData') ? undefined : EventDataSetFromJSON(json['eventData']),
        'additionalProps': !exists(json, 'additionalProps') ? undefined : AdditionalTimeSeriesPropertiesFromJSON(json['additionalProps']),
    };
}

export function TimeSeriesDataWithPropsToJSON(value?: TimeSeriesDataWithProps | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'eventData': EventDataSetToJSON(value.eventData),
        'additionalProps': AdditionalTimeSeriesPropertiesToJSON(value.additionalProps),
    };
}